combine user_storage and event_storage into single database module, separate imports (code style) _find_user_by_telegram_id change signature: single argument instead of two
