+++ bot/db/database.py
@@ -1,10 +1,17 @@
 from typing import Sequence
+
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
-from bot.db.db_engine import engine
-from bot.db.models import User
+from sqlalchemy.ext.asyncio import create_async_engine
+
+from bot.config_reader import env_config
+from bot.db.models import TelegramEvent, User
 from bot.user_dto import UserDTO
 
+engine = create_async_engine(
+    env_config.postgresql_url.get_secret_value(),
+    echo=True,
+)
 
 async def get_all_users() -> Sequence[User]:
     async with AsyncSession(bind=engine, autoflush=False) as async_session:
@@ -14,15 +21,13 @@ async def get_all_users() -> Sequence[User]:
 
 async def add_or_update_user(user_dto: UserDTO) -> None:
 
-    async def _find_user_by_telegram_id(
-        async_session: AsyncSession, telegram_id: int
-    ) -> User | None:
+    async def _find_user_by_telegram_id(telegram_id: int) -> User | None:
         statement = select(User).where(User.telegram_id == telegram_id)
         result = await async_session.execute(statement)
         return result.scalar_one_or_none()
 
     async with AsyncSession(bind=engine, autoflush=False) as async_session:
-        user = await _find_user_by_telegram_id(async_session, user_dto.telegram_id)
+        user = await _find_user_by_telegram_id(user_dto.telegram_id)
         if user is None:  # id пользователя нет в бд
             user = User(
                 telegram_id=user_dto.telegram_id,
@@ -36,4 +41,14 @@ async def add_or_update_user(user_dto: UserDTO) -> None:
             user.last_name = user_dto.last_name
             user.username = user_dto.username
 
+        await async_session.commit()
+
+
+async def create_telegram_event(telegram_id: int, payload: dict) -> None:
+    async with AsyncSession(bind=engine, autoflush=False) as async_session:
+        new_event = TelegramEvent(
+            telegram_id=telegram_id,
+            payload=payload,
+        )
+        async_session.add(new_event)
         await async_session.commit()
