+++ bot/handlers/__init__.py
@@ -1,5 +1,7 @@
 from typing import List
+
 from aiogram import Router
+
 from bot.handlers import (
     broadcast,
     buy_button,
