+++ bot/handlers/register_or_update_user.py
@@ -1,6 +1,7 @@
 from aiogram import Router
 from aiogram.types import Message
-from bot.db import user_storage
+
+from bot.db import database
 from bot.user_dto import UserDTO
 
 router = Router()
@@ -19,6 +20,6 @@ async def register_or_update_user(message: Message) -> None:
         username=user.username,
     )
 
-    await user_storage.add_or_update_user(user_dto)
+    await database.add_or_update_user(user_dto)
 
     print(f"User {user.id} is registered/updated in the db")
