+++ bot/middlewares/EventLoggerMiddleware.py
@@ -3,7 +3,7 @@ from typing import Any, Awaitable, Callable, cast, Dict
 from aiogram import BaseMiddleware
 from aiogram.types import TelegramObject, User
 
-from bot.db.event_storage import create_telegram_event
+from bot.db import database
 
 
 class EventLoggerMiddleware(BaseMiddleware):
@@ -15,5 +15,5 @@ class EventLoggerMiddleware(BaseMiddleware):
     ) -> Any:
         user = cast(User, data["event_from_user"])
         payload = event.model_dump(exclude_none=True)
-        await create_telegram_event(user.id, payload)
+        await database.create_telegram_event(user.id, payload)
         return await handler(event, data)
